backend/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0

# Development dependencies
pytest==7.4.0
black==23.7.0
flake8==6.1.0

